+++ ck2/CHANGES.md
@@ -1,7 +1,10 @@
 # Prototyping phase
 
 ## Master
+   - fixed major bug with inheritance (_base)
    - fixed a bug in "cm search" when automation is created without UID
+   - fixed a bug with @input.json
+   - fixed minior bugs in a core
 
 
 ## V0.7.4